SslCertificateDirectory generalized
On Windows `%userprofile%` is only interpreted in a few cases, depending on the environment it runs in. This does not include Visual Studio or PowerShell for example.

Using `Environment.GetFolderPath(Environment.SpecialFolder.UserProfile)` works always for Windows, Linux and Mac and C# supports forward slash for Windows.

Note: Linux was hardcoded to the root user before. Now it uses the current user.